# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CancelIn": {
        "properties": {
          "reason": {
            "maxLength": 500,
            "minLength": 12,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "CancelIn",
        "type": "object"
      },
...
      "DealIn": {
        "properties": {
          "buyer_entity_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buyer Entity Id"
          },
          "category": {
            "default": "apparel",
            "title": "Category",
            "type": "string"
          },
          "dispute_window_days": {
            "default": 7,
            "maximum": 90.0,
            "minimum": 0.0,
            "title": "Dispute Window Days",
            "type": "integer"
          },
          "milestones": {
            "items": {
              "$ref": "#/components/schemas/MilestoneIn"
            },
            "title": "Milestones",
            "type": "array"
          },
          "seller_entity_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seller Entity Id"
          },
          "seller_org_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seller Org Id"
          },
          "seller_org_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seller Org Slug"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "tolerance": {
            "additionalProperties": true,
            "title": "Tolerance",
            "type": "object"
          },
          "total_paise": {
            "exclusiveMinimum": 0.0,
            "title": "Total Paise",
            "type": "integer"
          }
        },
        "required": [
          "title",
          "total_paise",
          "milestones"
        ],
        "title": "DealIn",
        "type": "object"
      },
      "DealOut": {
        "properties": {
          "buyer_org": {
            "additionalProperties": true,
            "title": "Buyer Org",
            "type": "object"
          },
          "chain_deal_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chain Deal Id"
          },
          "chain_tx": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chain Tx"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dispute_window_days": {
            "title": "Dispute Window Days",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "milestones": {
            "items": {
              "$ref": "#/components/schemas/MilestoneOut"
            },
            "title": "Milestones",
            "type": "array"
          },
          "money": {
            "$ref": "#/components/schemas/MoneyOut"
          },
          "pricing_tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pricing Tier"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          },
          "risk_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Score"
          },
          "seller_org": {
            "additionalProperties": true,
            "title": "Seller Org",
            "type": "object"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "terms_hash": {
            "title": "Terms Hash",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "total_paise": {
            "title": "Total Paise",
            "type": "integer"
          },
          "viewer_side": {
            "enum": [
              "buyer",
              "seller"
            ],
            "title": "Viewer Side",
            "type": "string"
          }
        },
        "required": [
          "id",
          "reference",
          "title",
          "state",
          "total_paise",
          "money",
          "buyer_org",
          "seller_org",
          "terms_hash",
          "chain_deal_id",
          "chain_tx",
          "dispute_window_days",
          "risk_score",
          "pricing_tier",
          "milestones",
          "created_at",
          "viewer_side"
        ],
        "title": "DealOut",
        "type": "object"
      },
...
      "FundIn": {
        "properties": {
          "amount_paise": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Paise"
          }
        },
        "title": "FundIn",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MilestoneOut": {
        "properties": {
          "amount_paise": {
            "title": "Amount Paise",
            "type": "integer"
          },
          "attestation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attestation Id"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "has_evidence": {
            "default": false,
            "title": "Has Evidence",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "released_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Released At"
          },
          "seq": {
            "title": "Seq",
            "type": "integer"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "verification_condition": {
            "additionalProperties": true,
            "title": "Verification Condition",
            "type": "object"
          }
        },
        "required": [
          "id",
          "seq",
          "title",
          "amount_paise",
          "state",
          "verification_condition"
        ],
        "title": "MilestoneOut",
        "type": "object"
      },
...
      "SignTermsIn": {
        "properties": {
          "accept": {
            "default": true,
            "title": "Accept",
            "type": "boolean"
          }
        },
        "title": "SignTermsIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Programmable escrow for agentic commerce. The LLM never moves money: it writes a signed attestation, and a deterministic settlement engine reads it.",
    "title": "Aegis",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/deals": {
      "get": {
        "operationId": "list_deals_api_v1_deals_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DealOut"
                  },
                  "title": "Response List Deals Api V1 Deals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Deals",
        "tags": [
          "deals"
        ]
      },
      "post": {
        "operationId": "create_deal_api_v1_deals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DealIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Deal",
        "tags": [
          "deals"
        ]
      }
    },
...
    "/api/v1/deals/{deal_id}": {
      "get": {
        "operationId": "get_deal_api_v1_deals__deal_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "deal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Deal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Deal",
        "tags": [
          "deals"
        ]
      }
    },
    "/api/v1/deals/{deal_id}/cancel": {
      "post": {
        "operationId": "cancel_api_v1_deals__deal_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "deal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Deal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel",
        "tags": [
          "deals"
        ]
      }
    },
    "/api/v1/deals/{deal_id}/fund": {
      "post": {
        "operationId": "fund_api_v1_deals__deal_id__fund_post",
        "parameters": [
          {
            "in": "path",
            "name": "deal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Deal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FundIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fund",
        "tags": [
          "deals"
        ]
      }
    },
    "/api/v1/deals/{deal_id}/risk": {
      "get": {
        "operationId": "deal_risk_api_v1_deals__deal_id__risk_get",
        "parameters": [
          {
            "in": "path",
            "name": "deal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Deal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Deal Risk Api V1 Deals  Deal Id  Risk Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Deal Risk",
        "tags": [
          "deals"
        ]
      }
    },
    "/api/v1/deals/{deal_id}/sign-terms": {
      "post": {
        "operationId": "sign_terms_api_v1_deals__deal_id__sign_terms_post",
        "parameters": [
          {
            "in": "path",
            "name": "deal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Deal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignTermsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sign Terms",
        "tags": [
          "deals"
        ]
      }
    },
    "/api/v1/deals/{deal_id}/timeline": {
      "get": {
        "operationId": "timeline_api_v1_deals__deal_id__timeline_get",
        "parameters": [
          {
            "in": "path",
            "name": "deal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Deal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Timeline Api V1 Deals  Deal Id  Timeline Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Timeline",
        "tags": [
          "deals"
        ]
      }
    },
...
    "/api/v1/milestones/{milestone_id}": {
      "get": {
        "operationId": "get_milestone_api_v1_milestones__milestone_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "milestone_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Milestone Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MilestoneOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Milestone",
        "tags": [
          "deals"
        ]
      }
    },
...
    "/api/v1/milestones/{milestone_id}/start-verify": {
      "post": {
        "description": "Runs the verifier and, if it returns RELEASE, lets the engine authorise.",
        "operationId": "start_verify_api_v1_milestones__milestone_id__start_verify_post",
        "parameters": [
          {
            "in": "path",
            "name": "milestone_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Milestone Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Start Verify Api V1 Milestones  Milestone Id  Start Verify Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Verify",
        "tags": [
          "deals"
        ]
      }
    },
...
  }
}